Add HighTimer.hpp
Stopwatch is too complicated, that's why I screwed things up

+++ cpp/Stopwatch.cpp
@@ -6,13 +6,13 @@ Stopwatch::InternalData::InternalData()
     if (QueryPerformanceFrequency(&Frequency))
     {
         IsHighResolution = true;
-        TickFrequency = TicksPerSecond / Frequency.QuadPart;
+        TickFrequency = static_cast<long double>(TicksPerSecond) / Frequency.QuadPart;
     }
     else
     {
         IsHighResolution = false;
         Frequency.QuadPart = TicksPerSecond;
-        TickFrequency = 1;
+        TickFrequency = 1.0;
     }
 }
